refactor golang tests into one test package
will probably need to refactor this further, not sure I'll actually need the init facility... and some of the larger problems will cause timeouts for all the tests in the unit

but I can collect the sanity checks and only test larger values in the package's own main

most of the problems will answer within the default 30 second window, so tests for answers can still help maybe

+++ golang/p0003/solve.go
@@ -24,8 +24,8 @@ package p0003
 
 import "github.com/kevindamm/projecteuler/golang/util"
 
-func LargestPrimeFactor(n uint64) int {
-	var reduced = n
+func LargestPrimeFactor(n int) int {
+	var reduced = uint64(n)
 	for prime := range util.GeneratePrimesUntil(1 << 20) {
 		if prime > reduced>>1 {
 			return int(reduced)
